+++ spec/lib/aixm/feature/airport_spec.rb
@@ -240,8 +240,9 @@ describe AIXM::Feature::Airport do
           <geoLong>004.75216944E</geoLong>
           <valTrueBrg>165</valTrueBrg>
           <valMagBrg>166</valMagBrg>
-          <valElevTdz>147</valElevTdz>
+          <valElevTdz>145</valElevTdz>
           <uomElevTdz>FT</uomElevTdz>
+          <codeVfrPattern>E</codeVfrPattern>
           <txtRmk>forth remarks</txtRmk>
         </Rdn>
         <Rdd>
@@ -276,6 +277,9 @@ describe AIXM::Feature::Airport do
           <geoLong>004.75645556E</geoLong>
           <valTrueBrg>345</valTrueBrg>
           <valMagBrg>346</valMagBrg>
+          <valElevTdz>147</valElevTdz>
+          <uomElevTdz>FT</uomElevTdz>
+          <codeVfrPattern>L</codeVfrPattern>
           <txtRmk>back remarks</txtRmk>
         </Rdn>
         <Rdd>